Move exceptions into appropriate modules
