+++ mathesar/api/exceptions/database_exceptions/base_exceptions.py
@@ -1,0 +1,34 @@
+from rest_framework import status
+
+from mathesar.api.exceptions.error_codes import ErrorCodes
+from mathesar.api.exceptions.generic_exceptions.base_exceptions import MathesarAPIException
+
+
+class ProgrammingAPIException(MathesarAPIException):
+    error_code = ErrorCodes.ProgrammingError.value
+
+    def __init__(
+            self,
+            exception,
+            error_code=ErrorCodes.ProgrammingError.value,
+            message=None,
+            field=None,
+            details=None,
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR
+    ):
+        super().__init__(exception, error_code, message, field, details, status_code)
+
+
+class IntegrityAPIException(MathesarAPIException):
+    # Default message is not needed as the exception string provides enough details
+    error_code = ErrorCodes.NonClassifiedIntegrityError.value
+
+    def __init__(
+            self,
+            exception,
+            message=None,
+            field=None,
+            details=None,
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR
+    ):
+        super().__init__(exception, self.error_code, message, field, details, status_code)
