+++ mathesar/api/viewsets/columns.py
@@ -5,6 +5,8 @@ from rest_framework.exceptions import NotFound
 from rest_framework.response import Response
 from sqlalchemy.exc import ProgrammingError
 
+from mathesar.api.exceptions.database_exceptions import exceptions as database_exceptions, base_exceptions as base_database_exceptions
+from mathesar.api.exceptions.generic_exceptions import base_exceptions as base_api_exceptions
 from db.columns.exceptions import (
     DynamicDefaultWarning, InvalidDefaultError, InvalidTypeOptionError, InvalidTypeError,
 )
@@ -57,36 +59,36 @@ class ColumnViewSet(viewsets.ModelViewSet):
             except ProgrammingError as e:
                 if type(e.orig) == DuplicateColumn:
                     name = request.data['name']
-                    raise exceptions.DuplicateTableAPIException(
+                    raise database_exceptions.DuplicateTableAPIException(
                         e,
                         message=f'Column {name} already exists',
                         field='name',
                         status_code=status.HTTP_400_BAD_REQUEST
                     )
                 else:
-                    raise exceptions.ProgrammingAPIException(e)
+                    raise base_database_exceptions.ProgrammingAPIException(e)
             except TypeError as e:
-                raise exceptions.TypeErrorAPIException(
+                raise base_api_exceptions.TypeErrorAPIException(
                     e,
                     message="Unknown type_option passed",
                     status_code=status.HTTP_400_BAD_REQUEST
                 )
             except InvalidDefaultError as e:
-                raise exceptions.InvalidDefaultAPIException(
+                raise database_exceptions.InvalidDefaultAPIException(
                     e,
                     message=f'default "{request.data["default"]}" is invalid for type {request.data["type"]}',
                     status_code=status.HTTP_400_BAD_REQUEST
                 )
             except InvalidTypeOptionError as e:
                 type_options = request.data.get('type_options', '')
-                raise exceptions.InvalidTypeOptionAPIException(
+                raise database_exceptions.InvalidTypeOptionAPIException(
                     e,
                     message=f'parameter dict {type_options} is invalid for type {request.data["type"]}',
                     field="type_options",
                     status_code=status.HTTP_400_BAD_REQUEST
                 )
             except InvalidTypeError as e:
-                raise exceptions.InvalidTypeCastAPIException(
+                raise database_exceptions.InvalidTypeCastAPIException(
                     e, message='This type casting is invalid.',
                     status_code=status.HTTP_400_BAD_REQUEST
                 )
@@ -111,29 +113,29 @@ class ColumnViewSet(viewsets.ModelViewSet):
                 table.alter_column(column_instance._sa_column.column_index, serializer.validated_data)
             except ProgrammingError as e:
                 if type(e.orig) == UndefinedFunction:
-                    raise exceptions.UndefinedFunctionAPIException(
+                    raise database_exceptions.UndefinedFunctionAPIException(
                         e,
                         message='This type cast is not implemented',
                         status_code=status.HTTP_400_BAD_REQUEST
                     )
                 else:
-                    raise exceptions.ProgrammingAPIException(e, status_code=status.HTTP_400_BAD_REQUEST)
+                    raise base_database_exceptions.ProgrammingAPIException(e, status_code=status.HTTP_400_BAD_REQUEST)
             except IndexError as e:
                 raise exceptions.NotFoundAPIException(e)
             except TypeError as e:
-                raise exceptions.InvalidTypeOptionAPIException(
+                raise database_exceptions.InvalidTypeOptionAPIException(
                     e,
                     message="Unknown type_option passed",
                     status_code=status.HTTP_400_BAD_REQUEST
                 )
             except InvalidDefaultError as e:
-                raise exceptions.InvalidDefaultAPIException(
+                raise database_exceptions.InvalidDefaultAPIException(
                     e,
                     message=f'default "{request.data["default"]}" is invalid for this column',
                     status_code=status.HTTP_400_BAD_REQUEST
                 )
             except DynamicDefaultWarning as e:
-                raise exceptions.DynamicDefaultAPIException(
+                raise database_exceptions.DynamicDefaultAPIException(
                     e,
                     message='Changing type of columns with dynamically-generated defaults is not supported.'
                             'Delete or change the default first.',
@@ -141,19 +143,19 @@ class ColumnViewSet(viewsets.ModelViewSet):
                 )
             except InvalidTypeOptionError as e:
                 type_options = request.data.get('type_options', '')
-                raise exceptions.InvalidTypeOptionAPIException(
+                raise database_exceptions.InvalidTypeOptionAPIException(
                     e,
                     message=f'parameter dict {type_options} is invalid for type {request.data["type"]}',
                     status_code=status.HTTP_400_BAD_REQUEST
                 )
             except InvalidTypeError as e:
-                raise exceptions.InvalidTypeCastAPIException(
+                raise database_exceptions.InvalidTypeCastAPIException(
                     e,
                     message='This type casting is invalid.',
                     status_code=status.HTTP_400_BAD_REQUEST
                 )
             except Exception as e:
-                raise exceptions.MathesarAPIException(e)
+                raise base_api_exceptions.MathesarAPIException(e)
         serializer.update(column_instance, serializer.validated_model_fields)
         # Invalidate the cache as the underlying columns have changed
         out_serializer = ColumnSerializer(self.get_object())
