+++ mathesar/api/viewsets/constraints.py
@@ -4,6 +4,9 @@ from rest_framework.mixins import ListModelMixin, RetrieveModelMixin
 from rest_framework.response import Response
 from sqlalchemy.exc import ProgrammingError, IntegrityError
 
+import mathesar.api.exceptions.database_exceptions.base_exceptions
+import mathesar.api.exceptions.database_exceptions.exceptions
+import mathesar.api.exceptions.generic_exceptions.base_exceptions
 from mathesar.api.exceptions import exceptions
 from mathesar.api.pagination import DefaultLimitOffsetPagination
 from mathesar.api.serializers.constraints import ConstraintSerializer
@@ -35,21 +38,21 @@ class ConstraintViewSet(viewsets.GenericViewSet, ListModelMixin, RetrieveModelMi
             constraint = table.add_constraint(data['type'], data['columns'], name)
         except ProgrammingError as e:
             if type(e.orig) == DuplicateTable:
-                raise exceptions.DuplicateTableAPIException(
+                raise mathesar.api.exceptions.database_exceptions.exceptions.DuplicateTableAPIException(
                     e,
                     message='Relation with the same name already exists',
                     status_code=status.HTTP_400_BAD_REQUEST
                 )
             else:
-                raise exceptions.MathesarAPIException(e)
+                raise mathesar.api.exceptions.generic_exceptions.base_exceptions.MathesarAPIException(e)
         except IntegrityError as e:
             if type(e.orig) == UniqueViolation:
-                raise exceptions.UniqueViolationAPIException(
+                raise mathesar.api.exceptions.database_exceptions.exceptions.UniqueViolationAPIException(
                     e,
                     status_code=status.HTTP_400_BAD_REQUEST
                 )
             else:
-                raise exceptions.MathesarAPIException(e)
+                raise mathesar.api.exceptions.generic_exceptions.base_exceptions.MathesarAPIException(e)
 
         out_serializer = ConstraintSerializer(constraint, context={'request': request})
         return Response(out_serializer.data, status=status.HTTP_201_CREATED)
@@ -62,5 +65,5 @@ class ConstraintViewSet(viewsets.GenericViewSet, ListModelMixin, RetrieveModelMi
             if type(e.orig) == UndefinedObject:
                 raise exceptions.NotFoundAPIException(e)
             else:
-                raise exceptions.ProgrammingAPIException(e)
+                raise mathesar.api.exceptions.database_exceptions.base_exceptions.ProgrammingAPIException(e)
         return Response(status=status.HTTP_204_NO_CONTENT)
