+++ mathesar/api/viewsets/data_files.py
@@ -2,6 +2,9 @@ from rest_framework import status, viewsets
 from rest_framework.mixins import CreateModelMixin, ListModelMixin, RetrieveModelMixin
 from rest_framework.response import Response
 
+import mathesar.api.exceptions.data_import_exceptions.exceptions
+import mathesar.api.exceptions.database_exceptions.exceptions
+import mathesar.api.exceptions.generic_exceptions.base_exceptions
 from mathesar.api.exceptions.error_codes import ErrorCodes
 from mathesar.api.exceptions import exceptions
 from mathesar.errors import InvalidTableError
@@ -29,11 +32,11 @@ class DataFileViewSet(viewsets.GenericViewSet, ListModelMixin, RetrieveModelMixi
             serializer = DataFileSerializer(data_file, context={'request': request})
             return Response(serializer.data)
         else:
-            exception_body = exceptions.ErrorBody(
+            exception_body = mathesar.api.exceptions.generic_exceptions.base_exceptions.ErrorBody(
                 code=ErrorCodes.MethodNotAllowed.value,
                 message='Method "PATCH" allowed only for header.'
             )
-            raise exceptions.GenericAPIException(
+            raise mathesar.api.exceptions.generic_exceptions.base_exceptions.GenericAPIException(
                 [exception_body],
                 status.HTTP_405_METHOD_NOT_ALLOWED
             )
@@ -44,6 +47,6 @@ class DataFileViewSet(viewsets.GenericViewSet, ListModelMixin, RetrieveModelMixi
         try:
             datafile = create_datafile(serializer.validated_data)
         except InvalidTableError as e:
-            raise exceptions.InvalidTableAPIException(e, status_code=status.HTTP_400_BAD_REQUEST)
+            raise mathesar.api.exceptions.data_import_exceptions.exceptions.InvalidTableAPIException(e, status_code=status.HTTP_400_BAD_REQUEST)
         serializer = DataFileSerializer(datafile, context={'request': request})
         return Response(serializer.data, status=status.HTTP_201_CREATED)
